backend/requirements.txt
selenium==4.1.0
sqlite3==3.35.5
googlesearch-python==1.1.0
openai==0.27.0
beautifulsoup4==4.11.1
python-dotenv==0.19.2
